+++ src/main/java/com/jwplayer/jwplatform/exception/JWPlatformCallFailedException.java
@@ -1,0 +1,4 @@
+package com.jwplayer.jwplatform.exception;
+
+public class JWPlatformCallFailedException extends JWPlatformException {
+}