+++ src/main/java/com/jwplayer/jwplatform/exception/JWPlatformItemAlreadyExistsException.java
@@ -1,0 +1,4 @@
+package com.jwplayer.jwplatform.exception;
+
+public class JWPlatformItemAlreadyExistsException extends JWPlatformException {
+}