+++ src/main/java/com/jwplayer/jwplatform/exception/JWPlatformNoMethodException.java
@@ -1,0 +1,4 @@
+package com.jwplayer.jwplatform.exception;
+
+public class JWPlatformNoMethodException extends JWPlatformException {
+}