+++ src/main/java/com/jwplayer/jwplatform/exception/JWPlatformParameterEmptyException.java
@@ -1,0 +1,4 @@
+package com.jwplayer.jwplatform.exception;
+
+public class JWPlatformParameterEmptyException extends JWPlatformException {
+}