+++ src/main/java/com/jwplayer/jwplatform/exception/JWPlatformParameterEncodingException.java
@@ -1,0 +1,4 @@
+package com.jwplayer.jwplatform.exception;
+
+public class JWPlatformParameterEncodingException extends JWPlatformException {
+}