+++ src/main/java/com/jwplayer/jwplatform/exception/JWPlatformSignatureMissingException.java
@@ -1,0 +1,4 @@
+package com.jwplayer.jwplatform.exception;
+
+public class JWPlatformSignatureMissingException extends JWPlatformException {
+}