+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+* Add support for POST request endpoints.
+* Refactor exception handling such that exceptions are no longer nested classes of the factory. _This is a breaking change._
+
 ## 0.1.0
 * Initial version of the client.